RAC-153 refactor : usecase 이름 수정
